Rewrite package.json only if not configured for pre-git
Before, the install script always put all `config.pre-git` entries as
well as `scripts.commit` into the package.json regardless of whether
the user already did configure pre-git. This is not nice: If a user
wants to leave out some of those config entries, they are inserted
again later, e.g. on a later `npm install`. Thus with this commit the
entries are only written to package.json if there is no
`config.pre-git` key yet in it.

Also:
* Previously there was double code to fill `config.pre-git.commit-msg`.
  This duplication is now removed.
* The necessity to check whether the package.json was changed (and
  only write to it if so) became obsolete, because we now only write to
  it, if no `config.pre-git` is present yet.